Differentiate between primary and other actions
This is similar to the already existing Action.delete. Form's standard button is now primary.  But the button used on a table's query form is not, on the assumption that on many pages that will not be the default action.  If we had a standard Action.info or similar I would have used that -- maybe another day.