Every part can now have assets that are added to the assets of the style and included in the head.
This is particularly useful for bundling small pieces of javascript or css with the components that need them and thereby gets us closer to being able to write truly self contained "components".  As a proof of concept I did so for the tables javascript parts.  The naming takes care of deduplication of assets.  This can be witnessed in the new two table example (where the javascript is included just once).  Also hat-tip to Johan's cleanup of the examples they are much nicer to read now.