This changes the semantics of table.rows.
Idea: Introduce separate fields for the different stages (initial rows, sorted rows, filtered rows, visible rows).

Why is that good?  Because it means we bind / set each field once and the semantics does not depend on where we are
in the life-cycle of the table object (mutation is evil).

Indeed I ended up writing this patch because I had a bug trying to write a different patch, which happened because
I hadn't realized that self.paginator is used to rewrite self.rows to be just the visible rows.

As much as table.rows is considered to be a public API it changes the API (the the change in one of the tests).  Now
it didn't seem to have a clearly documented semantics so maybe this is fine.  At the moment I also haven't documented
the new fields.  Which is fine as long as table is primarily there as a way to display values and not to get them out
again.  But we could also consider documenting the fields better.

Testing: All tests still pass.  Also the table examples seem to still behave the same as before (but I only did
some superficial testing, without trying to exhaustively hit every interaction on every page).