No longer create a submit button by default.
But still allow the same concise definitions of the submit button as we did before.  How?  If there is something in actions__submit we by default assume it is a primary button.

As you have to define a post_handler to get anything meaningful anyway this should be silent to real clients.  The hopefully only place where this caused real breakage was some of our tests, as we lazily tend to only specify in the test what is strictly needed to get the test to run.