dunder field lookups for reverse 1-to-1 relationship: drop _set suffix
Fixes a regression introduced in 5.8.0.

Given an example:

```python3
class Parent(Model):
    name = CharField(max_length=255)

class Child(Model):
    parent = OneToOneField(Parent, on_delete=CASCADE)
    myfield = CharField(max_length=255)
```

Before 5.8.0, the dunder path to access the `Child.myfield` from `Parent`
in a Table was consistent with Django's behavior (`Parent.child` contains
the corresponding `Child` instance):

```python3
Table(auto__model=Parent, auto__include=["name", "child__myfield"])
```

Since 5.8.0, the above table raises an AssertioError `You can only
include fields that exist on the model`. In 5.8.0, the dunder path to
the field changed to `child_set__myfield`, which is inconsistent with
how the object is accessed on the model class in Django.

This commit brings back the pre-5.8 dunder path for one-to-one reverse
lookups.